requirements.txt
torch
torchvision
transformers
opencv-python
pandas
numpy
scikit-image
crewai
openai
matplotlib
python-dotenv
